==> code/00-about.qmd <==
# Contact
`pasta` is being developed by [Mark D. Robinson's lab](https://robinsonlabuzh.github.io/).

## About us
We are located in the [Department of Molecular Life Sciences](https://www.mls.uzh.ch/en.html) within the [Faculty of Science](https://www.mnf.uzh.ch/en.html) at the [University of Zurich](https://www.uzh.ch/en.html).

Research in our laboratory is focused on creating, comparing and understanding statistical methods and data science tools for processing and interpreting various types of genomic data.

## Team
Authors: [Martin Emons](https://orcid.org/0009-0000-5219-5311), [Samuel Gunz](https://orcid.org/0000-0002-8909-0932), [Malte Kuehl](https://orcid.org/0000-0003-4167-2498)(Python examples), [Izaskun Mallona](https://orcid.org/0000-0002-2853-7526), [Mark D. Robinson](https://orcid.org/0000-0002-3048-5518) \

We thank all members of the Robinsonlab for constructive feedback on this work. We thank Peiying Cai, Helena L. Crowell, Alice Driessen, Reinhard Furrer, Reto Gerber, Pierre-Luc Germain, Maruša Kodermann, Vladyslav Korobeynyk, Siyuan Luo, Izaskun Mallona, Giulia Moro, Emanuel Sonder, Jiayi Wang and David Wissel for their careful reading and input to the vignette.

## Contact
E-mail: `mark.robinson (at) mls.uzh.ch`

Address:
```
Department of Molecular Life Sciences
University of Zurich
Winterthurerstrasse 190
CH-8057 Zurich, Switzerland

Office Y11-G-05
Phone  +41 44 635 48 48
Fax    +41 44 635 68 68
```
